refactor(frontend): replace more barrel file imports
